[WGS_VIEWER][이원희][22/01/17-밑에 메뉴 사이즈 정상화-drawer에 넣음]

+++ lib/controller/left_chart_ctrl.dart
@@ -84,7 +84,7 @@ class ChartCtrl extends GetxController {
             if (TimeSelectCtrl.to.timeIdxList.length > Idx.value) {
               forfields[s][ii].add(WGSspot(
                   TimeSelectCtrl.to.timeIdxList[Idx.value],
-                  FilePickerCtrl.to.oesFD[s].avg[ii]));
+                  FilePickerCtrl.to.oesFD[s].avg[ii].toInt()));
             }
           }
         }
